+++ R/data_help_files.r
@@ -1,61 +1,51 @@
-#' @name story_keywords
-#' @title Story keywords
-#' @description Miscellaneous keywords for the episodes of Star Trek The Animated Series (TAS), The Original Series (TOS), and The Next Generation (TNG).
+#' @name story_ids_by_theme
+#' @title Story theme summary
+#' @description A table in which each theme is summarised according to the stories in which it is featured.
 #' @docType data
 #' @format Data frame
-#' @source The Star Trek Project. See http://http://www.thestartrekproject.net for more details.
+#' @source Theme Ontology. See http://www.themeontology.org for more details.
+NULL
+
+#' @name story_keywords
+#' @title Story keywords
+#' @description Miscellaneous keywords for the episodes of Star Trek The Original Series (tos), The Animated Series (tas), The Next Generation (tng), and Voyager (voy).
+#' @docType data
+#' @format Data frame
+#' @source Theme Ontology. See http://www.themeontology.org for more details.
 NULL
 
 #' @name story_settings
 #' @title Story settings
-#' @description Settings for the episodes of Star Trek The Animated Series (TAS), The Original Series (TOS), and The Next Generation (TNG).
+#' @description Settings for the episodes of Star Trek The Original Series (tos), The Animated Series (tas), The Next Generation (tng), and Voyager (voy).
 #' @docType data
 #' @format Data frame
-#' @source The Star Trek Project. See http://http://www.thestartrekproject.net for more details.
+#' @source Theme Ontology. See http://www.themeontology.org for more details.
 NULL
 
 #' @name story_metadata
 #' @title Story metadata
-#' @description Metadata for the episodes of Star Trek The Animated Series (TAS), The Original Series (TOS), and The Next Generation (TNG).
+#' @description Metadata for the episodes of Star Trek The Original Series (tos), The Animated Series (tas), The Next Generation (tng), and Voyager (voy).
 #' @docType data
 #' @format Data frame
-#' @source The Star Trek Project. See http://http://www.thestartrekproject.net for more details.
+#' @source Theme Ontology. See http://www.themeontology.org for more details.
 NULL
 
 
-#' @name themed_stories
+#' @name story_themes
 #' @title Themed stories
-#' @description Themes for the episodes of Star Trek The Animated Series (TAS), The Original Series (TOS), and The Next Generation (TNG).
+#' @description Themes for the episodes of Star Trek The Original Series (tos), The Animated Series (tas), The Next Generation (tng), and Voyager (voy).
 #' @docType data
 #' @format Data frame
-#' @source The Star Trek Project. See http://http://www.thestartrekproject.net for more details.
+#' @source Theme Ontology. See http://www.themeontology.org for more details.
 NULL
 
 
-#' @name theme_definitions
-#' @title Theme definitions
+#' @name theme_dict
+#' @title Theme dictionary
 #' @description A hierarchical structured vocabulary of defined themes in four domains: the human condition, society, alternate reality, and the pursuit of knowledge.
 #' @docType data
 #' @format Data frame
-#' @source The Star Trek Project. See http://http://www.thestartrekproject.net for more details.
+#' @source Theme Ontology. See http://www.themeontology.org for more details.
 NULL
-
-
-#' @name story_ids_by_theme
-#' @title Story theme summary
-#' @description A table in which each theme is summarised accoring to the stories in which it is featured.
-#' @docType data
-#' @format Data frame
-#' @source The Star Trek Project. See http://http://www.thestartrekproject.net for more details.
-NULL
-
-
-
-
-
-
-
-
-